ingredient-to-recipe/requirements.txt
streamlit==1.24.0
tiktoken==0.5.1
unstructured==0.10.16
langchain
python-dotenv==1.0.0
psycopg2-binary==2.9.6
altair==4.0.0
pgvector==0.2.3
urllib3==1.26.17
boto3==1.28.62
openai==0.28.1
transformers==4.34.0
torch==2.1.0
torchvision==0.16.0